EntityDynamic Access Update 1. ConnectionBuilder (List, Set, Map) update 2. DynamicEntityDao update, update() method updated 3. DynamicEntityDao query() api updated 4. CTO support clone() 5. Add cache control to TallyBookThymeleafViewResolver, using property key "thymeleaf.tallybook.extension.cache.enable"

+++ general/general-dynamic-dataservice/src/main/java/com/taoswork/tallybook/dynamic/dataservice/server/dto/request/EntityQueryRequest.java
@@ -64,7 +64,7 @@ public class EntityQueryRequest {
         return criteriaTransferObject.getFilterCriteriasCollection();
     }
 
-    public int getFirstResult() {
+    public long getFirstResult() {
         return criteriaTransferObject.getFirstResult();
     }
